Library/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
bcrypt==4.1.2